src/cleaner/requirements.txt
pandas
pika
flask
google-cloud-storage
prometheus_client
